Safari-611.1.10/LayoutTests/imported/w3c/web-platform-tests/html/rendering/widgets/button-layout/abspos-expected.txt

testtesttest

FAIL abspos button with auto width, non-auto left/right (rtl) assert_equals: offsetLeft expected 662 but got 102
FAIL abspos button with auto width, non-auto left/right (ltr) assert_equals: clientWidth expected 32 but got 592
